Make Serial::read and dependant functions 8-bit clean.

+++ cores/nRF5/Adafruit_TinyUSB_Core/Adafruit_USBD_CDC.cpp
@@ -96,7 +96,7 @@ int Adafruit_USBD_CDC::peek(void)
 
 int Adafruit_USBD_CDC::read(void)
 {
-  return (int) tud_cdc_read_char();
+  return tud_cdc_read_char();
 }
 
 void Adafruit_USBD_CDC::flush(void)
